Crate surge_ping

Macros

is_linux_icmp_socket

Structs

AsyncSocket

Client

If you want to pass the Client in the task, please wrap it with Arc: Arc<Client>. and can realize the simultaneous ping of multiple addresses when only one socket is created.

Config

Config is the packaging of various configurations of sockets. If you want to make some set_socket_opt and other modifications, please define and implement them in Config.

ConfigBuilder

Icmpv4Packet

Packet structure returned by ICMPv4.

Icmpv6Packet

Packet structure returned by ICMPv6.

PingIdentifier

PingSequence

Pinger

A Ping struct represents the state of one particular ping instance.

Enums

ICMP

IcmpPacket

Represents the ICMP reply packet.

SurgeError

An error resulting from a ping option-setting or send/receive operation.

Functions

ping

Shortcut method to ping address. NOTE: This function creates a new internal Client on each call, and so should not be used if making many target. Create a Client instead.